requirements.txt
python-nmap==0.7.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.0
asyncio==3.4.3
celery==5.3.5
redis==5.0.1
prometheus-client==0.18.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic[email]==2.5.2
cryptography==41.0.5
